+++ client/src/main/java/org/evosuite/symbolic/vm/wrappers/I_ParseInt.java
@@ -54,7 +54,7 @@ public final class I_ParseInt extends SymbolicFunction {
 				SymbolicHeap.$STRING_VALUE, conc_string, symb_string_ref,
 				conc_string);
 
-		long longValue = (long) conc_integer;
+		long longValue = conc_integer;
 
 		StringToIntegerCast parse_int_value = new StringToIntegerCast(
 				symb_string_value, longValue);
