+++ client/src/test/java/com/examples/with/different/packagename/concolic/TestCase47.java
@@ -588,7 +588,7 @@ public class TestCase47 {
 		charArray0[4] = charArray2[33];
 		int int1 = -4;
 		stemmer0.add(charArray0, int1);
-		stemmer0.add(charArray2, (int) charArray0[0]);
+		stemmer0.add(charArray2, charArray0[0]);
 
 	}
