+++ client/src/test/java/com/examples/with/different/packagename/solver/TestCaseCastIntToReal.java
@@ -23,7 +23,7 @@ public class TestCaseCastIntToReal {
 
 	public static boolean test(int int_x) {
 
-		double double_x = (double) int_x;
+		double double_x = int_x;
 		if (double_x == 0 && int_x != 0) {
 			return true;
 		} else {
